git-svn-id: https://open-delta.googlecode.com/svn/trunk@678 dd429c29-c832-59a0-e864-46e68368aa0b

+++ delta-editor/src/main/java/au/org/ala/delta/editor/slotfile/VOCharBaseDesc.java
@@ -204,7 +204,7 @@ public class VOCharBaseDesc extends VOImageHolderDesc {
 		return _fixedData.codedImplicit;
 	}
 
-	public void setCharType(short charType) {
+	public void setCharType(int charType) {
 		_fixedData.charType = charType;
 		setDirty();
 	}
